app: fix user nfc tag type

+++ app/app/src/main/java/de/stustanet/stustapay/model/User.kt
@@ -9,7 +9,7 @@ data class User(
     val display_name: String,
     val privileges: List<Privilege>,
     val description: String?,
-    val user_tag: Int? = null,
+    val user_tag: ULong? = null,
     val transport_account_id: Int? = null,
     val cashier_account_id: Int? = null,
 )
